requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
reportlab==4.0.7
psycopg2-binary==2.9.9
asyncpg==0.29.0
pandas==2.0.3
openpyxl==3.1.2
python-dotenv==1.0.0
PyJWT==2.8.0
